fix: ensure that state management works correctly again after structure changes
